Very hacky 2to3 import updates
Not working 100%, triple quotes in a couple of files are confusing
the very simple parsing. Also will either need to add 'as' clauses
to the import statements, or fix the main body of the code as well.
e.g. 'from Bio import SeqIO' --> 'from bio import seqio' works,
but need to change 'SeqIO.parse(...)' to 'seqio.parse(...)' too.
Or use 'from bio import seqio as SeqIO' but that seems ugly.