Lower-case module names hackery of Tutorial.tex
This is enough to get test_Tutorial.py to pass under Python 3
with the lower case module names. However, it does not update
all the examples or text in the tutorial, leaving it in a mixed
and somewhat confusing state from the point of view of using it
for documentation.

Perhaps making these lower case changes on the fly from within
test_Tutorial.py is wiser.