Return FASTQ and SFF quality scores as an array
Previously returned a list of integers, here we return a
unsigned byte integer array instead.

Note could do this on ABI and QUAL parser too, although
the later has a test case which causes an overflow.

This change in itself is a proof of principle, and to allow
checking of side effects. Step two would be to use this for
speeding up the parser...